refactor: rename files and relocate files
- rename apis.ts to utils.ts and user.mock.ts to mockApiHandler.mock.ts relocate utils.ts
- relocate utils.ts and mockApi.mock.ts
