+++ src/components/features/Login/EnrollForm/index.tsx
@@ -1,6 +1,6 @@
 import styled from '@emotion/styled';
 
-import { registerUser } from '@/api/user/api';
+import { registerUser } from '@/api/utils';
 import { Button } from '@/components/common/Button';
 import { UnderlineTextField } from '@/components/common/Form/Input/UnderlineTextField';
 import { Spacing } from '@/components/common/layouts/Spacing';
