+++ src/mocks/browser.ts
@@ -2,10 +2,10 @@ import { setupWorker } from 'msw';
 
 import { categoriesMockHandler } from '@/api/hooks/categories.mock';
 import { productsMockHandler } from '@/api/hooks/products.mock';
-import { userHandlers } from '@/api/user/user.mock';
+import { mockApiHandlers } from '@/api/mockApiHandler.mock';
 
 export const worker = setupWorker(
   ...categoriesMockHandler,
   ...productsMockHandler,
-  ...userHandlers,
+  ...mockApiHandlers,
 );
